Remove the quiet -d0 flag from b2
If enabled, travis will thinks the build is stalled because there is no
output for over ten minutes. If it is disabled, travis will not display
the whole log because it exceeds the log line limit.

Since incomplete logs are better than a failing build, i am reverting
that change.