Make backend constructor more generic
If you had a const char pointer and wanted to construct a APIEndpoint
with a Backend that takes a std::string, you could not do so directly.
You had to create the Backend first and then pass it to the APIEndpoint.

Now you can do this:
    dota2::WebEndpoint endpoint("key");
Instead of having to do this:
    dota2::WebEndpoint endpoint(dota2::WebAPIBackend("key"));