Make curl an imported library
We can create a new library for curl by setting its properties to the
values given to us by find_package. This has the following upsides:
- We can use namespaces (the double colon) to make sure we do not make a
typo in the library name. If the library name in target_link_libraries
does not contain a colon and is not a valid target, cmake will just link
against it with -l, which might fail at link time and not at generation
time.
- Using INTERFACE_INCLUDE_DIRECTORIES saves us the need to manually
include the include dirs for curl.